+++ src/test/__mocks__/common/hooks/useRecordControls.mock.ts
@@ -1,17 +1,11 @@
-export const saveRecord = jest.fn();
-export const discardRecord = jest.fn();
 export const fetchRecord = jest.fn();
-export const clearRecordState = jest.fn();
 export const fetchRecordAndSelectEntityValues = jest.fn();
 export const fetchExternalRecordForPreview = jest.fn();
 export const getRecordAndInitializeParsing = jest.fn();
 
 jest.mock('@/common/hooks/useRecordControls', () => ({
   useRecordControls: () => ({
-    saveRecord,
-    discardRecord,
     fetchRecord,
-    clearRecordState,
     fetchRecordAndSelectEntityValues,
     fetchExternalRecordForPreview,
     getRecordAndInitializeParsing,